requirements.txt
# FastAPI Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
onnxruntime>=1.16.0

# Database
supabase>=2.3.0
postgrest>=0.13.0

# Existing dependencies
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
rapidfuzz>=3.0.0
pillow>=10.0.0
python-Levenshtein>=0.20.0
huggingface-hub>=0.19.0

# Computer Vision and AI
opencv-python-headless>=4.8.0
torch>=2.6.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
datasets>=2.14.0
scipy>=1.11.0
ultralytics>=8.0.0
